Initializes other university bots in leaderboard
Moves the bot creation logic from the Game class to the
Leaderboard class, which makes more sense semantically.
These bots represent other universities that compete with
the player. Initializes the leaderboard with these bots
at the start of the game. Refactors Game class's initializeGame
method to reflect changes.
