+++ core/src/main/java/io/github/bilcitytycoon/Game.java
@@ -1,6 +1,7 @@
 package io.github.bilcitytycoon;
 
 import java.util.ArrayList;
+import java.util.Random;
 
 public class Game {
     private ArrayList<Building> allBuildings;
@@ -15,31 +16,9 @@ public class Game {
         this.leaderboard = new Leaderboard();
     }
 
-    public void createBots(){
-        Random rd = new Random();
-        //TODO: change university names
-        String[] universityNames = {
-            "Kanzi", "Delibas", "Undem", "Simsir", "Kafadar", "Yildirim", "Zeyno", "Zurna",
-            "Ardent", "Velbridge", "Nexora", "Stonewall", "Ironspire", "Thalium", "Ravenfall", "Mintora",
-            "Elarin", "Bronwyn", "Kairoth", "Solbridge", "Zentris", "Halveth", "Terranova", "Grantham",
-            "Obsidian", "Luneth", "Ashbourne", "Cryden", "Novaris", "Thornevale", "Valcrest", "Redspire",
-            "Averin", "Falkmore", "Vireon", "Pendwyn", "Brimley", "Zephyra", "Calchester", "Ironwell",
-            "Stormmere", "Crestmoor", "Marenth", "Eloria", "Westford", "Myndale", "Duskmoor", "Halbridge",
-            "Venthor", "Silvermark", "Brighthelm", "Norwyn", "Galderan", "Eastmere", "Draymar", "Borewyn",
-            "Trivent", "Runestone", "Wyrnfall", "Ashmor", "Cindergate", "Starwyn", "Frostmoor", "Veriton",
-            "Greymark", "Tarnhelm", "Quanthelm", "Blackmere", "Ebonridge", "Cloudmere", "Highvale", "Bronzefield",
-            "Glenhaven", "Thornford", "Solwyn", "Vandale", "Mistbrook", "Ismere", "Redwyn", "Goldbarrow",
-            "Newwyn", "Maveth", "Stonewyn", "Westoria", "Rynmere", "Netherford", "Vortexa", "Hawkmere",
-            "Graveth", "Skyreach", "Cypresson", "Kryden", "Lorewyn", "Zenmere", "Altmore", "Brighthall",
-            "Vistaryn", "Oakspire", "Stormreach"
-        };
 
-        for(int i=0;i<universityNames.length;i++){
-            allUniversities.add(new OtherUniversity(universityNames[i],rd.nextInt(1000,5001),rd.nextInt(20,100),this.leaderboard));
-        }
-    }
-    public void initilaizeGame(){
-        createBots();
+    public void initializeGame(){
+        //leaderboard initilaize
         //TODO
     }
 }
