+++ core/src/main/java/io/github/bilcitytycoon/Map.java
@@ -1,4 +1,5 @@
 package io.github.bilcitytycoon;
 
 public class Map {
+    //TODO
 }
